requirements.txt
httpx>=0.27.0
pyyaml>=6.0
openai>=1.50.0
python-frontmatter>=1.1.0
pytest>=8.0.0
